dependency heap graph building moved to dnWalker.Symbolic - can be reused by other components
