+++ dnWalker.Symbolic/Heap/Graphs/DependencyGroup.cs
@@ -1,13 +1,11 @@
-﻿using dnWalker.Symbolic.Heap;
-
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace dnWalker.TestGenerator.Symbolic.Heap
+namespace dnWalker.Symbolic.Heap.Graphs
 {
     public class DependencyGroup : IReadOnlyCollection<IReadOnlyHeapNode>
     {
@@ -16,7 +14,7 @@ namespace dnWalker.TestGenerator.Symbolic.Heap
         public DependencyGroup(IEnumerable<IReadOnlyHeapNode> heapNodes)
         {
             _heapNodes = new List<IReadOnlyHeapNode>(heapNodes);
-            
+
             // nodes within one group have proper ordering
             _heapNodes.Sort(static (n1, n2) => n1.Location.Value.CompareTo(n2.Location.Value));
         }
