+++ dnWalker.Symbolic/dnWalker.Symbolic.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="dnlib" Version="3.5.0" />
+    <PackageReference Include="QuikGraph" Version="2.5.0" />
   </ItemGroup>
 
   <ItemGroup>
